Major changes to correlate with the changes to the interface: 1) All cursor pairs now converted to ranges.  Other misc interface    changes ported. 2) Almost complete rewrite of smart cursors and ranges.    This is incomplete and crashes, but Cullmann asked for it to be    committed so that it can be worked on at akademy. 3) KateSmartManager is the new way I'm trying to efficiently handle    updating smart cursors and ranges when something changes.  This is where    most of the breakage is at the moment, i think. 4) Some work on the renderer - needs more 5) KateEditInfo is the new way I'm working on to store document change    history in more detail.  It's similar to the old way but I want it to be    more comprehensive.  Also, we'll be able to take a reference to the    document at any point in time, and translate a cursor or range from then    to any other point in time
svn path=/trunk/KDE/kdelibs/kate/; revision=455076

+++ part/part/katelinerange.cpp
@@ -95,7 +95,7 @@ bool KateLineLayout::startsInvisibleBlock() const
   if (!isValid() || virtualLine() == 0)
     return false;
 
-  return virtualLine() - 1 != m_doc->getVirtualLine(line() - 1);
+  return (virtualLine() - 1) != (int)m_doc->getVirtualLine(line() - 1);
 }
 
 int KateLineLayout::shiftX() const
